backend/requirements.txt
asgiref==3.2.10
Django==3.1
django-cors-headers==3.7.0
djangorestframework==3.12.4
Owlready2==0.33
pytz==2021.1
sqlparse==0.4.1
psycopg2==2.9.1
sql-metadata==2.2.2